feat(Shared With You): extract carousel view and hero view from cells so we can reuse it for shared with you cells
